bug: move _CustomSource to private file
